requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.2
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
filelock==3.12.2
importlib-metadata==6.4.0
iniconfig==2.0.0
Markdown==3.4.3
packaging==23.1
platformdirs==3.8.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
virtualenv==20.23.1
zipp==3.15.0
